Fixed "Pushing the same view controller instance more than once is not supported"
Problem: I created the view controller of the drawer item that navigates
to at the same time I create the drawer item. This leads to a problem
when user clicks on the item a second time - I am pushing the same view
controller again. I worked around it by tracking the last indexPath
pushed. If they're the same, don't navigate again. The correct approach
should be instaniating the view controller at the time of navigation.
This commit fixes that.